refactor(sftp): de-duplicate transfer ops and inline SFTPOperation
